添加了搜索特定课程的平均分数值的相关方法

+++ src/main/java/com/dd/ai_smart_course/dto/ScoreDTO.java
@@ -2,11 +2,11 @@ package com.dd.ai_smart_course.dto;
 
 import lombok.Data;
 
-import java.util.List;
+import java.math.BigDecimal;
 
 @Data
 public class ScoreDTO {
     private Integer studentId;
     private Integer courseId;
-    private Integer finalScore;  //课程最终分数
+    private BigDecimal finalScore;  //课程最终分数
 }
